+++ pkg/providers/appsec/testdata/ase-apr/appsec-variables.tf
@@ -1,0 +1,42 @@
+variable "group_name" {
+  type    = string
+  default = ""
+}
+
+variable "contract_id" {
+  type    = string
+  default = ""
+}
+
+variable "name" {
+  type    = string
+  default = "TFDEMO"
+}
+
+variable "description" {
+  type    = string
+  default = <<EOT
+A security config for
+demo
+EOT
+}
+
+variable "hostnames" {
+  type    = list(string)
+  default = ["test.akamai.com"]
+}
+
+variable "emails" {
+  type    = list(string)
+  default = ["noreply@example.org"]
+}
+
+variable "activation_note" {
+  type    = string
+  default = "Activated by Terraform"
+}
+
+variable "network" {
+  type    = string
+  default = "STAGING"
+}