+++ pkg/providers/appsec/testdata/ase-apr/modules/activate-security/versions.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_providers {
+    akamai = {
+      source  = "akamai/akamai"
+      version = ">= 8.0.0"
+    }
+  }
+  required_version = ">= 1.0"
+
+}