+++ pkg/providers/appsec/testdata/ase-apr/modules/security/policies.tf
@@ -1,0 +1,7 @@
+resource "akamai_appsec_security_policy" "default_policy" {
+  config_id              = local.config_id
+  default_settings       = true
+  security_policy_name   = "Default Policy"
+  security_policy_prefix = "ASE1"
+}
+